requirements.txt
# this is not read by setup. It is here solely for RTD and user interest
requests
ipython
numpy
scipy
pandas
pebble
rdkit-pypi
rdkit-to-params
molecular-rectifier
smallworld-api  # optional
sqlitedict  # optional
nglview  # optional

# `pymol2` is conda only and optional
# `pyrosetta` has non-standard installation, see docs and `fragmenstein.igor.pyrosetta_import` for mock for RTD
